+++ benchmarks/charts/CHANGELOG.md
@@ -1,5 +1,13 @@
 # charts-benchmark
 
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@6.4.0
+  - @inglorious/charts@10.0.0
+
 ## 0.1.16
 
 ### Patch Changes