+++ examples/apps/web-element-size/CHANGELOG.md
@@ -1,5 +1,13 @@
 # web-element-size
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@6.4.0
+  - @inglorious/ui@7.0.0
+
 ## 0.0.2
 
 ### Patch Changes