+++ examples/apps/web-finance-dashboard/CHANGELOG.md
@@ -1,5 +1,14 @@
 # web-finance-dashboard
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@6.4.0
+  - @inglorious/charts@10.0.0
+  - @inglorious/ui@7.0.0
+
 ## 0.1.7
 
 ### Patch Changes