+++ examples/apps/web-todomvc-ace/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-todomvc-ace
 
+## 0.1.26
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@6.4.0
+
 ## 0.1.25
 
 ### Patch Changes