+++ packages/charts/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inglorious/charts
 
+## 10.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@6.4.0
+
 ## 9.0.0
 
 ### Patch Changes